Support pluggable captcha systems
This change turns the captcha support into an API that allows for
other captcha systems to be used.  The existing captcha system (based
on figlet) is still supported and its API continues to be respected
(i.e., `use-captcha?`, `audible-captcha?` etc.).

This change introduces a new keyword parameter for `pastiche`
(`captcha-api`).  It can be either `#f` or a thunk that returns a
one-argument procedure.  The one-argument procedure will be given
messages (symbols) as arguments:

* `render`: request the SXML to be added to the paste form page.  It
  is assumed to be suitable to be added to a table (in SXML, it should
  return a list of lists -- see `tabularize` in `pastiche.scm`.

* `valid?`: check whether the user answer for the captcha is valid.
  The user answer will be in the request variable as specified in the
  form field returned by `render`.

* `consume!`: in case of captchas that must be invalidated once they
  are used, this message will be given when it is time to consume a
  captcha. `(void)` can be a suitable response to this message in case
  the captcha system doesn't need to invalidate captchas as they are
  used.

* `define-pages`: in case the captcha system requires new pages, they
  can be defined upon receiving this message. `(void)` can be a
  suitable response to this message in case the captcha system doesn't
  need to define extra pages.